+++ doc/changelog.md
@@ -1,6 +1,11 @@
 
 X-Plane Obj Library
 ---------------------------------------------------------------------------
+#### Next
+
+- **Fixed:** Printing custom name of the LIGHT_PARAM.
+
+---------------------------------------------------------------------------
 #### 0.4.1-beta (24.08.2017)
 
 - **Added:** Printing error if texture name contains illegal symbols.